requirements.txt
Django==2.1
django-appconf==1.0.3
django-imagekit==4.0.2
django-mdeditor==0.1.16
django-password-reset==2.0
Markdown==3.1.1
pilkit==2.0
Pillow==6.1.0
Pygments==2.4.2
pytz==2019.2
six==1.13.0
